+++ src/django_multisite_plus/cms_urls.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from django.conf.urls import include
 from djangocms_multisite.urlresolvers import cms_multisite_url