+++ src/django_multisite_plus/conf.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals, absolute_import
 import os
 
 from aldryn_addons.utils import boolean_ish