+++ src/django_multisite_plus/migrations/0001_initial.py
@@ -22,6 +22,7 @@ class Migration(migrations.Migration):
                         primary_key=True,
                         serialize=False,
                         to="sites.Site",
+                        on_delete=models.CASCADE,
                     ),
                 ),
                 (