+++ tests/test_fastrouter_lookup.py
@@ -1,0 +1,1 @@
+from django_multisite_plus import fastrouter_lookup